+++ tests/run.lua
@@ -10,6 +10,7 @@ package.path = here .. "/?.lua;" .. package.path
 
 local lu = require("luaunit")
 require("test_add_code_files")
+require("test_crossref_anx")
 require("test_filetree")
 require("test_r_syntax")